requirements.txt
Django~=5.0.6
djangorestframework
django-filter